requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
cffi==1.15.1
charset-normalizer==2.1.1
Django==4.1.1
django-mathfilters==1.0.0
idna==3.4
Pillow==9.2.0
pycparser==2.21
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
